cleanup(weight-transfer): consolidate load graph
